Crate datafusion_physical_expr

Re-exports

pub use aggregate::AggregateExpr;

pub use equivalence::EquivalenceProperties;

pub use equivalence::EquivalentClass;

pub use planner::create_physical_expr;

pub use expressions::BooleanQueryExpr;

pub use expressions::boolean_query;

pub use expressions::boolean_query_with_cnf;

pub use expressions::BooleanQueryEvalFunc;

pub use utils::expr_list_eq_any_order;

pub use utils::expr_list_eq_strict_order;

pub use utils::normalize_expr_with_equivalence_properties;

pub use utils::normalize_out_expr_with_alias_schema;

pub use utils::normalize_sort_expr_with_equivalence_properties;

pub use utils::sort_expr_list_eq_strict_order;

pub use utils::split_conjunction;

Modules

aggregate

array_expressions

Array expressions

conditional_expressions

crypto_expressions

Crypto expressions

datetime_expressions

DateTime expressions

equivalence

execution_props

expressions

Defines physical expressions that can evaluated at runtime during query execution

from_slice

A trait to define from_slice functions for arrow types

functions

Declaration of built-in (scalar) functions. This module contains built-in functions’ enumeration and metadata.

hash_utils

Functionality used both on logical and physical plans

math_expressions

Math expressions

planner

regex_expressions

Regex expressions

rewrite

string_expressions

String expressions

struct_expressions

Struct expressions

type_coercion

Type coercion rules for functions with multiple valid signatures

udf

UDF support

unicode_expressions

Unicode expressions

utils

var_provider

Variable provider

window

Macros

analysis_expect

binary_array_op

The binary_array_op macro includes types that extend beyond the primitive, such as Utf8 strings.

binary_array_op_dyn_scalar

The binary_array_op_dyn_scalar macro includes types that extend beyond the primitive, such as Utf8 strings.

Structs

AnalysisContext

The shared context used during the analysis of an expression. Includes the boundaries for all known columns.

ExprBoundaries

Represents the boundaries of the resulting value from a physical expression, if it were to be an expression, if it were to be evaluated.

PhysicalSortExpr

Represents Sort operation for a column in a RecordBatch

ScalarFunctionExpr

Physical expression of a scalar function

Traits

PhysicalExpr

Expression that can be evaluated against a RecordBatch A Physical expression knows its type, nullability and how to evaluate itself.