Trait datafusion_physical_expr::aggregate::AggregateExpr

pub trait AggregateExpr: Send + Sync + Debug {
    // Required methods
    fn as_any(&self) -> &dyn Any;
    fn field(&self) -> Result<Field>;
    fn create_accumulator(&self) -> Result<Box<dyn Accumulator>>;
    fn state_fields(&self) -> Result<Vec<Field>>;
    fn expressions(&self) -> Vec<Arc<dyn PhysicalExpr>> ;

    // Provided methods
    fn name(&self) -> &str { ... }
    fn row_accumulator_supported(&self) -> bool { ... }
    fn supports_bounded_execution(&self) -> bool { ... }
    fn create_row_accumulator(
        &self,
        _start_index: usize
    ) -> Result<Box<dyn RowAccumulator>> { ... }
    fn reverse_expr(&self) -> Option<Arc<dyn AggregateExpr>> { ... }
    fn create_sliding_accumulator(&self) -> Result<Box<dyn Accumulator>> { ... }
}

An aggregate expression that:

knows its resulting field

knows how to create its accumulator

knows its accumulator’s state’s field

knows the expressions from whose its accumulator will receive values

Required Methods

fn as_any(&self) -> &dyn Any

Returns the aggregate expression as Any so that it can be downcast to a specific implementation.

fn field(&self) -> Result<Field>

the field of the final result of this aggregation.

fn create_accumulator(&self) -> Result<Box<dyn Accumulator>>

the accumulator used to accumulate values from the expressions. the accumulator expects the same number of arguments as expressions and must return states with the same description as state_fields

fn state_fields(&self) -> Result<Vec<Field>>

the fields that encapsulate the Accumulator’s state the number of fields here equals the number of states that the accumulator contains

fn expressions(&self) -> Vec<Arc<dyn PhysicalExpr>>

expressions that are passed to the Accumulator. Single-column aggregations such as sum return a single value, others (e.g. cov) return many.

Provided Methods

fn name(&self) -> &str

Human readable name such as "MIN(c2)". The default implementation returns placeholder text.

fn row_accumulator_supported(&self) -> bool

If the aggregate expression is supported by row format

fn supports_bounded_execution(&self) -> bool

Specifies whether this aggregate function can run using bounded memory. Any accumulator returning “true” needs to implement retract_batch.

fn create_row_accumulator( &self, _start_index: usize ) -> Result<Box<dyn RowAccumulator>>

RowAccumulator to access/update row-based aggregation state in-place. Currently, row accumulator only supports states of fixed-sized type.

We recommend implementing RowAccumulator along with the standard Accumulator, when its state is of fixed size, as RowAccumulator is more memory efficient and CPU-friendly.

fn reverse_expr(&self) -> Option<Arc<dyn AggregateExpr>>

Construct an expression that calculates the aggregate in reverse. Typically the “reverse” expression is itself (e.g. SUM, COUNT). For aggregates that do not support calculation in reverse, returns None (which is the default value).

fn create_sliding_accumulator(&self) -> Result<Box<dyn Accumulator>>

Creates accumulator implementation that supports retract