Module datafusion_physical_expr::functions

Declaration of built-in (scalar) functions. This module contains built-in functions’ enumeration and metadata.

Generally, a function has:

a signature

a return type, that is a function of the incoming argument’s types

the computation, that must accept each valid signature

Signature: see Signature

Return type: a function (arg_types) -> return_type. E.g. for sqrt, (f32) -> f32, (f64) -> f64.

This module also supports coercion to improve user experience: if an argument i32 is passed to a function that supports f64, the argument is automatically is coerced to f64.

Functions

create_physical_expr

Create a physical (function) expression. This function errors when args’ can’t be coerced to a valid argument type of the function.

create_physical_fun

Create a physical scalar function.

make_scalar_function

decorates a function to handle ScalarValues by converting them to arrays before calling the function and vice-versa after evaluation.