Module datafusion_physical_expr::utils

Functions

expr_list_eq_any_order

Compare the two expr lists are equal no matter the order. For example two InListExpr can be considered to be equals no matter the order:

expr_list_eq_strict_order

Strictly compare the two expr lists are equal in the given order.

normalize_expr_with_equivalence_properties

normalize_out_expr_with_alias_schema

Normalize the output expressions based on Alias Map and SchemaRef.

normalize_sort_expr_with_equivalence_properties

ordering_satisfy

Checks whether given ordering requirements are satisfied by provided PhysicalSortExprs.

ordering_satisfy_concrete

sort_expr_list_eq_strict_order

Strictly compare the two sort expr lists in the given order.

split_conjunction

Assume the predicate is in the form of CNF, split the predicate to a Vec of PhysicalExprs.